# optimized toMidi function by 7-10 times