i made toMidi() function much faster but, at the cost of not having numbers and stringified numbers as input.
so these test cases will fail:
```ts
expect(Midi.toMidi("60")).toBe(60);
expect(Midi.toMidi(60)).toBe(60);
```
benchmarks were made on a [separate repository](https://github.com/eye-wave/fast-note-parser)
```
my-parser: [MIN] 5.63 ms
my-parser: [MAX] 25.37 ms
my-parser: [AVG] 15.82 ms

tonal-midi: [MIN] 140.65 ms
tonal-midi: [MAX] 176.28 ms
tonal-midi: [AVG] 150.44 ms
```

also i'm sorry for that messy commit diff, i thought it was formatted and linted beforehand