@danigb 2023-09-10
Hi, thanks for the PR. There are a couple of issues:
- It changes the behaviour of the function, and I want to keep compatibility. All previous test should pass, including the ones with numbers. Also, previously it accept any number of sharps or flats. Can you implement that too?
- As you mention, it seems that you accidentally formatted some not related code. Can you revert those files? (maybe by calling `npm run format`)

🙌