@eye-wave 2023-09-12
i tried to add those edge cases, but the performance was worse than your implementation. i won't bother you any longer then with this pr